refactor: 패키지 구조 변경 #7
- common/domain의 내용을 하나의 Dir로 합침.
