+++ src/main/java/site/praytogether/pray_together/exception/GlobalExceptionHandler.java
@@ -1,11 +1,11 @@
-package site.praytogether.pray_together.common.exception;
+package site.praytogether.pray_together.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import site.praytogether.pray_together.common.exception.spec.ExceptionSpec;
+import site.praytogether.pray_together.exception.spec.ExceptionSpec;
 
 @Slf4j
 @RestControllerAdvice
